Implemented parseUnaryExpression function

+++ src/ASTNumberExprNode.cpp
@@ -4,7 +4,7 @@
 
 #include "include/ASTNodes/ASTNumberExprNode.h"
 
-ASTNumberExprNode::ASTNumberExprNode(float value, std::string type) {
+ASTNumberExprNode::ASTNumberExprNode(double value, std::string type) {
     this->value = value;
     this->type = type;
 }
